+++ mrmustard/physics/tensornetwork.py
@@ -12,65 +12,276 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-This module contains functions for handling tensor networks in MrMustard
-"""
+""" Tensor class for constructing circuits out of components."""
 
-from typing import Optional
+from __future__ import annotations
 
-import matplotlib.pyplot as plt
-import networkx as nx
-import numpy as np
-
-from mrmustard import settings
-from mrmustard.lab.abstract import State
-from mrmustard.physics.circuit_part import CircuitPart, Wire
+from dataclasses import dataclass, field
+from typing import List, Optional
 from mrmustard.math import Math
-from mrmustard.physics.fock import autocutoffs
 
 math = Math()
 
 
-def connect(wire1: Wire, wire2: Wire, dim: Optional[int] = None):
-    r"""Connects a wire of this CircuitPart to another Wire (of the same or another CircuitPart).
+@dataclass
+class Wire:
+    r"""Represents a wire in a tensor network.
+
+    Args:
+        id: A numerical identifier for this wire.
+        mode: The mode represented by this wire.
+        is_input: Whether this wire is an input to a tensor or an output.
+        is_ket: Whether this wire is on the ket or on the bra side ??.
+        connection_id: A numerical identifier for the contraction involving this wire, or ``None``
+            if this wire is not contracted.
+        dimension: The dimension of this wire.
+        connected_to: The identifier of the tensor connected to this wire, or ``None`` if this wire
+            is not connected.
+
+    """
+    id: int
+    mode: int
+    is_input: bool
+    is_ket: bool
+
+    def __post_init__(self):
+        self._contraction_id: int | None = None
+        self._dimension: int | None = None
+        self._connected_to: int | None = None
+
+
+@dataclass
+class WireGroup:
+    r"""A group of wires in a tensor network.
+
+    Args:
+        ket: ??
+        bra: ??
+
+    """
+    ket: dict = field(default_factory=dict)
+    bra: dict = field(default_factory=dict)
+
+
+class Tensor:
+    r"""A tensor in a tensor network.
+
+    Args:
+        name (str): The name of this tensor.
+        input_legs_ket (List[int]): The indeces labelling the input legs on the bra side.
+        output_legs_ket (List[int]): The indeces labelling the output legs on the ket side.
+        input_legs_bra (List[int]): The indeces labelling the input legs on the bra side.
+        output_legs_bra (List[int]): The indeces labelling the output legs on the ket side.
+    """
+    _id_counter: int = 0  # to give a unique id to all Tensors and Wires
+    _repr_markdown_ = None  # otherwise it takes over the repr due to mro
+
+    def __init__(
+        self,
+        name: str,
+        input_legs_ket: list[int] = [],
+        output_legs_ket: list[int] = [],
+        input_legs_bra: list[int] = [],
+        output_legs_bra: list[int] = [],
+    ) -> None:
+        # set unique self.id and name
+        self.id: int = self._new_id()
+        self.name: str = name + "_" + str(self.id)
+
+        # initialize ket and bra wire dicts
+        self._in = WireGroup()
+        self._out = WireGroup()
+
+        # initialize wires by updating the ket and bra dicts
+        for mode in input_legs_ket:
+            self._in.ket |= {mode: Wire(self._new_id(), mode, True, True)}
+        for mode in output_legs_ket:
+            self._out.ket |= {mode: Wire(self._new_id(), mode, False, True)}
+        for mode in input_legs_bra:
+            self._in.bra |= {mode: Wire(self._new_id(), mode, True, False)}
+        for mode in output_legs_bra:
+            self._out.bra |= {mode: Wire(self._new_id(), mode, False, False)}
+
+    @property
+    def wires(self) -> List[Wire]:
+        r"""Returns a list of all wires in this tensor.
+        The order is MM default: [ket_out, ket_in, bra_out, bra_in].
+        However, minimize reliance on ordering in favour of ids and direction/type.
+        """
+        return (
+            list(self.output.ket.values())
+            + list(self.input.ket.values())
+            + list(self.output.bra.values())
+            + list(self.input.bra.values())
+        )
+
+    @property
+    def contraction_ids(self) -> list[int]:
+        r"""Returns a list of all contraction_ids in this Tensor."""
+        return [wire.connection_id for wire in self.wires]
+
+    def wire(self, id: int) -> Wire:
+        r"""Returns the wire with the given id."""
+        for wire in self.wires:
+            if wire.id == id:
+                return wire
+
+    def _new_id(self) -> int:
+        id = Tensor._id_counter
+        Tensor._id_counter += 1
+        return id
+
+    @property
+    def input(self):
+        return self._in
+
+    @property
+    def output(self):
+        return self._out
+
+    @property
+    def modes(self) -> list[int]:
+        r"""For backward compatibility. Don't overuse.
+        It returns a list of modes for this Tensor, unless it's ambiguous."""
+        if self.modes_in == self.modes_out:  # transformation on same modes
+            return self.modes_in
+        elif len(self.modes_in) == 0:  # state
+            return self.modes_out
+        elif len(self.modes_out) == 0:  # measurement
+            return self.modes_in
+        else:
+            raise ValueError("modes are ambiguous for this Tensor.")
+
+    @property
+    def modes_in(self) -> set[int]:
+        "Returns the set of input modes that are used by this Tensor."
+        return set(self.input.ket | self.input.bra)
+
+    @property
+    def modes_out(self) -> set[int]:
+        "Returns the set of output modes that are used by this Tensor."
+        return set(self.output.ket | self.output.bra)
+
+    @property
+    def all_modes(self) -> set[int]:
+        "Returns a set of all the modes spanned by this Tensor."
+        return self.modes_out + self.modes_in
+
+    @property
+    def adjoint(self) -> AdjointView:
+        r"""Returns the adjoint view of this Tensor (with new ids). That is, ket <-> bra."""
+        return AdjointView(self)
+
+    @property
+    def dual(self) -> DualView:
+        r"""Returns the dual view of this Tensor (with new ids). That is, in <-> out."""
+        return DualView(self)
+
+    @property
+    def view(self) -> TensorView:
+        r"""Returns a view of this Tensor with new ids."""
+        return TensorView(self)
+
+
+class TensorView(Tensor):
+    r"""Base class for Tensor views. It remaps the ids of the original Tensor."""
+
+    def __init__(self, circuit_part):
+        self._original = circuit_part  # MM object
+        super().__init__(
+            self._original.name,
+            self._original.output.ket.keys(),
+            self._original.input.ket.keys(),
+            self._original.output.bra.keys(),
+            self._original.input.bra.keys(),
+        )
+
+        # note that for the zip we are relying on wire ordering, which is not ideal
+        self._id_map = {
+            wire.id: wire_orig.id for wire, wire_orig in zip(self.wires, self._original.wires)
+        }
+
+    def _original_id(self, id):
+        return self._id_map[id]
+
+    def __getattr__(self, attr):
+        orig_attr = self._original.__getattribute__(attr)
+        if callable(orig_attr):
+
+            def method(*args, **kwargs):
+                kwargs["id"] = self._original_id(kwargs["id"])
+                return orig_attr(*args, **kwargs)
+
+            return method
+        return orig_attr
+
+
+class DualView(TensorView):
+    r"""Dual view of a Tensor. It is used to implement the dual.
+    It swaps the input and output wires of a Tensor.
+    """
+
+    def __new__(cls, circuit_part: Tensor):
+        "makes sure that DualView(DualView(circuit_part)) == TensorView(circuit_part)"
+        if isinstance(circuit_part, DualView):
+            return circuit_part._original
+        return super().__new__(cls)
+
+    @property
+    def input(self):
+        return self._original.output
+
+    @property
+    def output(self):
+        return self._original.input
+
+    @property
+    def dual(self):
+        return self._original.view
+
+
+class AdjointView(TensorView):
+    r"""Adjoint view of a Tensor. It is used to implement the adjoint.
+    It swaps the ket and bra wires of a Tensor.
+    """
+
+    def __new__(cls, circuit_part: Tensor):
+        "makes sure that AdjointView(AdjointView(circuit_part)) == TensorView(circuit_part)"
+        if isinstance(circuit_part, AdjointView):
+            return circuit_part._original
+        return super().__new__(cls)
+
+    @property
+    def input(self):  # swaps ket and bra
+        return WireGroup(ket=self._original.input.bra, bra=self._original.input.ket)
+
+    @property
+    def output(self):  # swaps ket and bra
+        return WireGroup(ket=self._original.output.bra, bra=self._original.output.ket)
+
+    @property
+    def adjoint(self):
+        return self._original.view
+
+
+def connect(wire1: Wire, wire2: Wire, dim: int):
+    r"""Connects two wires in a tensor network.
     Arguments:
         wire1: the first wire
         wire2: the second wire
-        dim (optional int): set the dimension of the contraction (if None, we use autocutoff
-                    for States or we use the largest of the dimension of the two wires)
+        dim (int): the dimension of the contraction
     """
-    wire1.connected_to = wire2
-    wire2.connected_to = wire1
-    # set dimension to max of the two wires if exists
-    if not dim:
-        dim1 = get_dimension(wire1) if isinstance(wire1.owner, State) else None
-        dim2 = get_dimension(wire2) if isinstance(wire2.owner, State) else None
-    dim = dim or max(dim1, dim2, key=lambda x: x or 0)
-    if dim is None:
-        raise ValueError("Dimension cannot be inferred from wires. Please set it manually.")
-    wire1.dimension = wire2.dimension = dim
-    wire2.contraction_id = wire1.contraction_id
+    wire1._connected_to = wire2
+    wire2._connected_to = wire1
+
+    wire1._dimension = dim
+    wire2._dimension = dim
+
+    wire1._connection_id = 1
+    wire2._connection_id = 1
 
 
-# TODO: revisit when we have Bargmann by default
-def get_dimension(wire: Wire, probability: Optional[float] = None) -> Optional[int]:
-    r"""Returns the dimension of a wire (fock cutoff)
-    Arguments:
-        wire (Wire): the wire
-    Returns:
-        (int) the dimension of the wire if it is a State, None otherwise
-    """
-    if isinstance(wire.owner, State):
-        i = wire.owner.modes.index(wire.mode)
-        j = i + len(wire.owner.modes)
-        cov = wire.owner.cov
-        sub_cov = np.array([[cov[i, i], cov[i, j]], [cov[j, i], cov[j, j]]])
-        means = wire.owner.means
-        sub_means = np.array([means[i], means[j]])
-        return autocutoffs(sub_cov, sub_means, probability or settings.AUTOCUTOFF_PROBABILITY)[0]
-
-
-def contract(tensors: list[CircuitPart], default_dim):
+def contract(tensors: list[Tensor], default_dim):
     r"""Contract a list of tensors.
     Arguments:
         tensors: the tensors to contract
@@ -81,108 +292,7 @@ def contract(tensors: list[CircuitPart], default_dim):
     opt_einsum_args = []
     for t in tensors:
         for w in t.wires:
-            w.dimension = w.dimension or default_dim
+            w._dimension = w._dimension or default_dim
         opt_einsum_args.append(t.fock)
-        opt_einsum_args.append([w.contraction_id for w in t.wires])
+        opt_einsum_args.append([w._connection_id for w in t.wires])
     return math.einsum(*opt_einsum_args)
-
-
-### DRAWING ###
-
-
-def draw(
-    tensors: list[CircuitPart],
-    show_direction=True,
-    show_modes=True,
-    show_types=False,
-    figsize=(6, 6),
-):
-    r"""
-    Draws a tensor network using networkx.
-    Arguments:
-        tensors (sequence of CircuitPart): the tensors to draw
-        show_direction (bool): whether to show the direction in the node labels (default: True)
-        show_modes (bool): whether to show the modes in the node labels (default: True)
-        show_types (bool): whether to show the types in the node labels (default: False)
-    """
-    G = nx.Graph()
-
-    ### NODES AND EDGES ###
-    # Add fake edges between the wires of the same tensor
-    for tensor in tensors:
-        for i, wire1 in enumerate(tensor.wires):
-            for wire2 in tensor.wires[i + 1 :]:
-                G.add_edge(wire1.id, wire2.id, color="grey", style="dashed")
-    # Add real edges between contracted wires
-    done = []
-    for tensor in tensors:
-        for wire in tensor.wires:
-            if wire.connected_to and wire.id not in done and wire.connected_to not in done:
-                G.add_edge(wire.id, wire.connected_to.id, color="red", style="solid")
-                done.append(wire.id)
-                done.append(wire.connected_to.id)
-
-    ### LABELS ###
-    node_labels = {}
-    edge_labels = {}
-    for tensor in tensors:
-        for wire in tensor.wires:
-            label_parts = (
-                [wire.direction] * show_direction
-                + [wire.type] * show_types
-                + [str(wire.mode)] * show_modes
-            )
-            node_labels[wire.id] = "_".join(label_parts)
-            if wire.connected_to:
-                edge_labels[(wire.id, wire.connected_to.id)] = str(wire.dimension)
-
-    # positions for all nodes
-    pos = nx.kamada_kawai_layout(G)
-
-    # positions for all nodes
-    edge_label_pos = {k: [v[0], v[1]] for k, v in pos.items()}
-
-    plt.figure(figsize=figsize)
-    edges = G.edges(data=True)
-    colors = [d["color"] for u, v, d in edges]
-
-    # Draw the nodes and edges
-    nx.draw(G, pos, edge_color=colors, labels=node_labels)
-    nx.draw_networkx_edge_labels(G, edge_label_pos, edge_labels=edge_labels)
-
-    # Draw bounding boxes around the tensors
-    for tensor in tensors:
-        x_min, y_min, x_max, y_max = _get_bounding_box(pos, [w.id for w in tensor.wires])
-        _draw_bounding_box(x_min, y_min, x_max, y_max)
-        # add tensor name on top of the bounding box
-        plt.text(
-            x_min + 0.03,
-            y_max + 0.0,
-            tensor.short_name,
-            horizontalalignment="center",
-            verticalalignment="bottom",
-        )
-
-    plt.title("Tensor Network")
-    plt.show()
-
-
-def _get_bounding_box(pos, node_ids, offset=0.03):
-    node_positions = [pos[node_id] for node_id in node_ids]
-
-    x_coordinates = [pos[0] for pos in node_positions]
-    y_coordinates = [pos[1] for pos in node_positions]
-
-    x_min, x_max = min(x_coordinates), max(x_coordinates)
-    y_min, y_max = min(y_coordinates), max(y_coordinates)
-    return (
-        x_min - 1.6 * offset,
-        y_min - offset,
-        x_max + 1.6 * offset,
-        y_max + offset,
-    )  # NOTE: 1.6 otherwise boxes aren't square 😵‍💫??
-
-
-def _draw_bounding_box(x_min, y_min, x_max, y_max):
-    box = plt.Rectangle((x_min, y_min), x_max - x_min, y_max - y_min, fill=None, color="blue")
-    plt.gca().add_patch(box)
